CSV field validation, exceptions for error status in open_url
- CSVField moved from costume_loader_pkg/schemas.py to api/extra_fields.py, added docstring
- optional raise_on_error_status parameter added to open_url in requests.py
- updates in costume_filter
